--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -1,10 +1,10 @@
-module github.com/fixme_my_friend/hw10_program_optimization
+module github.com/guamoko995/hw-otus/hw10_program_optimization
 
 go 1.16
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/mailru/easyjson v0.7.7
 	github.com/stretchr/testify v1.7.0
-	gopkg.in/yaml.v2 v2.4.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )